memobride-backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
bcrypt==4.0.1
Pillow==10.0.1
python-multipart==0.0.6
openai==1.3.0
langchain==0.0.346
python-dateutil==2.8.2